[feat] Creation of this flake
Basic functioning docker image; it can be loaded
or specified in virtualisation NixOS module.

On start, the container will call wgcf register and
wgcf generate to setup new/default warp account and
wireguard profile;

Then a SOCKS5 section is inserted into the profile
and used by wireproxy; the instance listen on port
25555 for traffic and 25556 for monitoring; the port
setup is not parameterized for now.

Healthcheck is unimplemented as well.